# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PlayerState = { [key: string]: unknown };

export interface Player {
  id: string;
  recoveryCode: string;
  state: PlayerState;
}

export type SavePlayerRequestState = { [key: string]: unknown };

export interface SavePlayerRequest {
  id: string;
  recoveryCode: string;
  state: SavePlayerRequestState;
}