# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerState } from './playerState';

export interface Player {
  id: string;
  recoveryCode: string;
  state: PlayerState;
}