# lib/api-zod/src/generated/types/savePlayerRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SavePlayerRequestState } from './savePlayerRequestState';

export interface SavePlayerRequest {
  id: string;
  recoveryCode: string;
  state: SavePlayerRequestState;
}